requirements.txt
# Core
fastapi==0.115.11
uvicorn==0.34.0
httpx==0.28.1
python-dotenv==1.0.1
pydantic==2.10.6
apscheduler==3.10.4

# YouTube
youtube-transcript-api==1.1.0

# LLM — Gemini (기본)
google-genai

# LLM — 선택적 (사용할 provider만 설치)
# openai
# anthropic

# 유틸리티
python-dateutil>=2.8.2